+++ media/colors.lua
@@ -1,9 +1,0 @@
---[[ xCT+ TBC Anniversary Classic
-     Author: paradosi-Dreamscythe
-     MIT License ]]
-
-local ADDON_NAME, addon = ...
-
--- Shorten my handle
-local x = addon.engine
-